+++ lib/rubocop/cop/vicenzo_cops.rb
@@ -2,3 +2,4 @@
 
 require_relative 'vicenzo/rspec/nested_let_redefinition'
 require_relative 'vicenzo/rspec/nested_subject_redefinition'
+require_relative 'vicenzo/rails/enum_inclusion_of_validation'