+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Initial release;
 - Add RoboCop::Cop::Vicenzo::RSpec::NestedLetRedefinition #1;
 - Add RoboCop::Cop::Vicenzo::RSpec::NestedSubjectRedefinition #2;
+- Add RuboCop::Cop::Vicenzo::Rails::EnumInclusionOfValidation #3;